REF/TST: method-specific files for rename, reset_index (#32227)
