+++ pandas/tests/series/test_alter_axes.py
@@ -3,7 +3,7 @@ from datetime import datetime
 import numpy as np
 import pytest
 
-from pandas import DataFrame, Index, MultiIndex, RangeIndex, Series
+from pandas import Index, MultiIndex, Series
 import pandas._testing as tm
 
 
@@ -31,62 +31,6 @@ class TestSeriesAlterAxes:
 
     # Renaming
 
-    def test_rename(self, datetime_series):
-        ts = datetime_series
-        renamer = lambda x: x.strftime("%Y%m%d")
-        renamed = ts.rename(renamer)
-        assert renamed.index[0] == renamer(ts.index[0])
-
-        # dict
-        rename_dict = dict(zip(ts.index, renamed.index))
-        renamed2 = ts.rename(rename_dict)
-        tm.assert_series_equal(renamed, renamed2)
-
-        # partial dict
-        s = Series(np.arange(4), index=["a", "b", "c", "d"], dtype="int64")
-        renamed = s.rename({"b": "foo", "d": "bar"})
-        tm.assert_index_equal(renamed.index, Index(["a", "foo", "c", "bar"]))
-
-        # index with name
-        renamer = Series(
-            np.arange(4), index=Index(["a", "b", "c", "d"], name="name"), dtype="int64"
-        )
-        renamed = renamer.rename({})
-        assert renamed.index.name == renamer.index.name
-
-    def test_rename_by_series(self):
-        s = Series(range(5), name="foo")
-        renamer = Series({1: 10, 2: 20})
-        result = s.rename(renamer)
-        expected = Series(range(5), index=[0, 10, 20, 3, 4], name="foo")
-        tm.assert_series_equal(result, expected)
-
-    def test_rename_set_name(self):
-        s = Series(range(4), index=list("abcd"))
-        for name in ["foo", 123, 123.0, datetime(2001, 11, 11), ("foo",)]:
-            result = s.rename(name)
-            assert result.name == name
-            tm.assert_numpy_array_equal(result.index.values, s.index.values)
-            assert s.name is None
-
-    def test_rename_set_name_inplace(self):
-        s = Series(range(3), index=list("abc"))
-        for name in ["foo", 123, 123.0, datetime(2001, 11, 11), ("foo",)]:
-            s.rename(name, inplace=True)
-            assert s.name == name
-
-            exp = np.array(["a", "b", "c"], dtype=np.object_)
-            tm.assert_numpy_array_equal(s.index.values, exp)
-
-    def test_rename_axis_supported(self):
-        # Supporting axis for compatibility, detailed in GH-18589
-        s = Series(range(5))
-        s.rename({}, axis=0)
-        s.rename({}, axis="index")
-        # TODO: clean up shared index validation
-        # with pytest.raises(ValueError, match="No axis named 5"):
-        #     s.rename({}, axis=5)
-
     def test_set_name_attribute(self):
         s = Series([1, 2, 3])
         s2 = Series([1, 2, 3], name="bar")
@@ -103,107 +47,12 @@ class TestSeriesAlterAxes:
         assert s.name is None
         assert s is not s2
 
-    def test_rename_inplace(self, datetime_series):
-        renamer = lambda x: x.strftime("%Y%m%d")
-        expected = renamer(datetime_series.index[0])
-
-        datetime_series.rename(renamer, inplace=True)
-        assert datetime_series.index[0] == expected
-
     def test_set_index_makes_timeseries(self):
         idx = tm.makeDateIndex(10)
 
         s = Series(range(10))
         s.index = idx
         assert s.index.is_all_dates
-
-    def test_reset_index(self):
-        df = tm.makeDataFrame()[:5]
-        ser = df.stack()
-        ser.index.names = ["hash", "category"]
-
-        ser.name = "value"
-        df = ser.reset_index()
-        assert "value" in df
-
-        df = ser.reset_index(name="value2")
-        assert "value2" in df
-
-        # check inplace
-        s = ser.reset_index(drop=True)
-        s2 = ser
-        s2.reset_index(drop=True, inplace=True)
-        tm.assert_series_equal(s, s2)
-
-        # level
-        index = MultiIndex(
-            levels=[["bar"], ["one", "two", "three"], [0, 1]],
-            codes=[[0, 0, 0, 0, 0, 0], [0, 1, 2, 0, 1, 2], [0, 1, 0, 1, 0, 1]],
-        )
-        s = Series(np.random.randn(6), index=index)
-        rs = s.reset_index(level=1)
-        assert len(rs.columns) == 2
-
-        rs = s.reset_index(level=[0, 2], drop=True)
-        tm.assert_index_equal(rs.index, Index(index.get_level_values(1)))
-        assert isinstance(rs, Series)
-
-    def test_reset_index_name(self):
-        s = Series([1, 2, 3], index=Index(range(3), name="x"))
-        assert s.reset_index().index.name is None
-        assert s.reset_index(drop=True).index.name is None
-
-    def test_reset_index_level(self):
-        df = DataFrame([[1, 2, 3], [4, 5, 6]], columns=["A", "B", "C"])
-
-        for levels in ["A", "B"], [0, 1]:
-            # With MultiIndex
-            s = df.set_index(["A", "B"])["C"]
-
-            result = s.reset_index(level=levels[0])
-            tm.assert_frame_equal(result, df.set_index("B"))
-
-            result = s.reset_index(level=levels[:1])
-            tm.assert_frame_equal(result, df.set_index("B"))
-
-            result = s.reset_index(level=levels)
-            tm.assert_frame_equal(result, df)
-
-            result = df.set_index(["A", "B"]).reset_index(level=levels, drop=True)
-            tm.assert_frame_equal(result, df[["C"]])
-
-            with pytest.raises(KeyError, match="Level E "):
-                s.reset_index(level=["A", "E"])
-
-            # With single-level Index
-            s = df.set_index("A")["B"]
-
-            result = s.reset_index(level=levels[0])
-            tm.assert_frame_equal(result, df[["A", "B"]])
-
-            result = s.reset_index(level=levels[:1])
-            tm.assert_frame_equal(result, df[["A", "B"]])
-
-            result = s.reset_index(level=levels[0], drop=True)
-            tm.assert_series_equal(result, df["B"])
-
-            with pytest.raises(IndexError, match="Too many levels"):
-                s.reset_index(level=[0, 1, 2])
-
-        # Check that .reset_index([],drop=True) doesn't fail
-        result = Series(range(4)).reset_index([], drop=True)
-        expected = Series(range(4))
-        tm.assert_series_equal(result, expected)
-
-    def test_reset_index_range(self):
-        # GH 12071
-        s = Series(range(2), name="A", dtype="int64")
-        series_result = s.reset_index()
-        assert isinstance(series_result.index, RangeIndex)
-        series_expected = DataFrame(
-            [[0, 0], [1, 1]], columns=["index", "A"], index=RangeIndex(stop=2)
-        )
-        tm.assert_frame_equal(series_result, series_expected)
 
     def test_reorder_levels(self):
         index = MultiIndex(
@@ -268,25 +117,6 @@ class TestSeriesAlterAxes:
         expected = Series([1, 2, 3], index=expected_index)
         tm.assert_series_equal(result, expected)
 
-    def test_rename_with_custom_indexer(self):
-        # GH 27814
-        class MyIndexer:
-            pass
-
-        ix = MyIndexer()
-        s = Series([1, 2, 3]).rename(ix)
-        assert s.name is ix
-
-    def test_rename_with_custom_indexer_inplace(self):
-        # GH 27814
-        class MyIndexer:
-            pass
-
-        ix = MyIndexer()
-        s = Series([1, 2, 3])
-        s.rename(ix, inplace=True)
-        assert s.name is ix
-
     def test_set_axis_inplace_axes(self, axis_series):
         # GH14636
         ser = Series(np.arange(4), index=[1, 3, 5, 7], dtype="int64")
@@ -322,21 +152,6 @@ class TestSeriesAlterAxes:
         for axis in [2, "foo"]:
             with pytest.raises(ValueError, match="No axis named"):
                 s.set_axis(list("abcd"), axis=axis, inplace=False)
-
-    def test_reset_index_drop_errors(self):
-        #  GH 20925
-
-        # KeyError raised for series index when passed level name is missing
-        s = Series(range(4))
-        with pytest.raises(KeyError, match="does not match index name"):
-            s.reset_index("wrong", drop=True)
-        with pytest.raises(KeyError, match="does not match index name"):
-            s.reset_index("wrong")
-
-        # KeyError raised for series when level to be dropped is missing
-        s = Series(range(4), index=MultiIndex.from_product([[1, 2]] * 2))
-        with pytest.raises(KeyError, match="not found"):
-            s.reset_index("wrong", drop=True)
 
     def test_droplevel(self):
         # GH20342
